+++ tests/xerbla_test.c
@@ -18,24 +18,24 @@
 #include "semicolon_lapack_auxiliary.h"
 
 /* Global state (replaces Fortran COMMON blocks) */
-int    xerbla_infot  = 0;
-int    xerbla_nout   = 0;
-int    xerbla_ok     = 1;
-int    xerbla_lerr   = 0;
+INT    xerbla_infot  = 0;
+INT    xerbla_nout   = 0;
+INT    xerbla_ok     = 1;
+INT    xerbla_lerr   = 0;
 char   xerbla_srnamt[33] = "";
 
-void xerbla(const char* srname, int info) {
+void xerbla(const char* srname, INT info) {
     /* When not in error-testing mode, behave like production xerbla */
     if (xerbla_srnamt[0] == '\0') {
-        fprintf(stderr, " ** On entry to %s parameter number %d had an illegal value\n",
-                srname, info);
+        fprintf(stderr, " ** On entry to %s parameter number %lld had an illegal value\n",
+                srname, (long long)info);
         return;
     }
 
     xerbla_lerr = 1;
     if (info != xerbla_infot) {
-        fprintf(stderr, " *** XERBLA was called with INFO = %d instead of %d ***\n",
-                info, xerbla_infot);
+        fprintf(stderr, " *** XERBLA was called with INFO = %lld instead of %lld ***\n",
+                (long long)info, (long long)xerbla_infot);
         fprintf(stderr, " *** routine = %s ***\n", srname);
         xerbla_ok = 0;
     }
